+++ libs/cardano-ledger-api/CHANGELOG.md
@@ -1,8 +1,10 @@
 # Version history for `cardano-ledger-api`
 
-## 1.2.2.0
+## 1.3.0.0
 
 * Addition of `Cardano.Ledger.Api.Tx.Cert`
+* Removal of `VoterRole` in favor of `Voter`
+* Removal of `cgVoterRolesL` and `cgVoterRoles` as no longer needed.
 
 ## 1.2.1.0